Move launcher functions inside trollflow2 package for testability
